Rewrite state manager
